clippy

+++ littlefs-rust-core/tests/test_badblocks.rs
@@ -93,7 +93,7 @@ fn test_badblocks_single(
             let size = NAMEMULT as u32;
             for _j in 0..(i * FILEMULT) {
                 let n = lfs_file_write(lfs, file, &buffer[..size as usize]);
-                assert_eq!(n, Ok(size as u32));
+                assert_eq!(n, Ok(size));
             }
 
             assert_ok(lfs_file_close(lfs, file));
@@ -136,7 +136,7 @@ fn test_badblocks_single(
             for _j in 0..(i * FILEMULT) {
                 let mut rbuffer = [0u8; 1024];
                 let n = lfs_file_read(lfs, file, &mut rbuffer[..size as usize]);
-                assert_eq!(n, Ok(size as u32));
+                assert_eq!(n, Ok(size));
                 assert_eq!(&rbuffer[..size as usize], &buffer[..size as usize]);
             }
 
@@ -313,7 +313,7 @@ fn badblocks_create_dirs_and_files(lfs: &mut Lfs) {
         let size = NAMEMULT as u32;
         for _j in 0..(i * FILEMULT) {
             let n = lfs_file_write(lfs, file, &buffer[..size as usize]);
-            assert_eq!(n, Ok(size as u32));
+            assert_eq!(n, Ok(size));
         }
 
         assert_ok(lfs_file_close(lfs, file));
@@ -354,7 +354,7 @@ fn badblocks_verify_dirs_and_files(lfs: &mut Lfs) {
         for _j in 0..(i * FILEMULT) {
             let mut rbuffer = [0u8; 1024];
             let n = lfs_file_read(lfs, file, &mut rbuffer[..size as usize]);
-            assert_eq!(n, Ok(size as u32));
+            assert_eq!(n, Ok(size));
             assert_eq!(&rbuffer[..size as usize], &buffer[..size as usize]);
         }
